+++ src/coinbase/webhook.ts
@@ -1,6 +1,5 @@
 import { Webhook as WebhookModel, WebhookEventType, WebhookEventFilter } from "../client/api";
 import { Coinbase } from "./coinbase";
-import { InternalError } from "./errors";
 
 /**
  * A representation of a Webhook,
@@ -13,11 +12,11 @@ export class Webhook {
    * Initializes a new Webhook object.
    *
    * @param model - The underlying Webhook object.
-   * @throws {InternalError} If the model is not provided.
+   * @throws {Error} If the model is not provided.
    */
   private constructor(model: WebhookModel) {
     if (!model) {
-      throw new InternalError("Webhook model cannot be empty");
+      throw new Error("Webhook model cannot be empty");
     }
     this.model = model;
   }