+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- Improved error mesasges for `InternalError`
+
 ## [0.1.1] - 2024-08-27
 
 - Fixed a bug where `listHistoricalBalances` method was parsing conventional ETH balances instead of atomic units